requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.3
idna==3.2
lxml==4.6.3
pyaes==1.6.1
pyasn1==0.4.8
python-dateutil==2.8.2
python-dotenv==0.18.0
requests==2.26.0
rsa==4.7.2
six==1.16.0
Telethon==1.23.0
urllib3==1.26.6
webdavclient3==3.14.5
